Structure change
Changed directory structure for Data and FitResults (moved to a common
place for all methods).

Included demo data directories.

Fit results are now saved automatically in a new FitResults directory
inside working directory.

Now with full Nifti file support. Fit results are also saved to .nii
maps.

Changed where the Sf table is saved (for SledPike models). It is now
saved inside the Protocol.mat file, since it is specific for each
protocol.

Changed the button to build Sf table to the Protocol panel and removed
the Load Sf table button (no longer a separate file).

Reworked the buld Sf function to add angles and offsets to the current
protocol automatically, so that the interpolation can be more accurate.
Points outside the range of current values are set to 0.25*min and
1.75*max of bothe angles and offsets. Two points are added between each
pair of points after that.

Set default protocols to empty ones, to make it clearer for users that
they should set their own to match to their data. Included
'DemoProtocol.mat' for each method instead.


Former-commit-id: c76273e789c8a37f8ad2dcc78bb8733ff0172a47